tsunagu_project/requirements.txt
django
pillow
django-shortuuidfield
